# swaggerJson/swagger-ledger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/bill": {
      "post": {
        "operationId": "BillController_add",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer JWT Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBillDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "path": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "data": {
                      "$ref": "#/components/schemas/CreateBillRes"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "新增",
        "tags": [
          "账本-账单"
        ]
      },
      "get": {
        "operationId": "BillController_list",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "页码，从 1 开始",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": true,
            "in": "query",
            "description": "每页数量，默认为 10",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "账单名称",
            "schema": {
              "example": "账单名称",
              "type": "string"
            }
          },
          {
            "name": "ledgerId",
            "required": true,
            "in": "query",
            "description": "账本id",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "classifyId",
            "required": false,
            "in": "query",
            "description": "账单分类id",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "consumerId",
            "required": false,
            "in": "query",
            "description": "消费人id",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "billDate",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2025-01-01",
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2025-01-01",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2025-01-31",
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer JWT Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "页码，从 1 开始",
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "default": 1,
              "example": 1
            }
          },
          {
            "description": "每页数量，默认为 10",
            "name": "size",
            "in": "query",
            "required": true,
            "schema": {
              "default": 10,
              "example": 10
            }
          },
          {
            "description": "账单名称",
            "required": false,
            "name": "name",
            "in": "query",
            "schema": {
              "example": "账单名称"
            }
          },
          {
            "description": "账本id",
            "name": "ledgerId",
            "in": "query",
            "required": true,
            "schema": {
              "example": 1
            }
          },
          {
            "description": "账单分类id",
            "required": false,
            "name": "classifyId",
            "in": "query",
            "schema": {
              "example": 1
            }
          },
          {
            "description": "消费人id",
            "required": false,
            "name": "consumerId",
            "in": "query",
            "schema": {
              "example": 1
            }
          },
          {
            "required": false,
            "name": "billDate",
            "in": "query",
            "schema": {
              "example": "2025-01-01"
            }
          },
          {
            "required": false,
            "name": "startDate",
            "in": "query",
            "schema": {
              "example": "2025-01-01"
            }
          },
          {
            "required": false,
            "name": "endDate",
            "in": "query",
            "schema": {
              "example": "2025-01-31"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "path": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ListBillRes"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "分页列表",
        "tags": [
          "账本-账单"
        ]
      }
    },
    "/bill/{id}": {
      "put": {
        "operationId": "BillController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "数据id",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer JWT Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBillDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "path": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "data": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "更新",
        "tags": [
          "账本-账单"
        ]
      },
      "get": {
        "operationId": "BillController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "数据id",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer JWT Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "path": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "data": {
                      "$ref": "#/components/schemas/DetailBillRes"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          },
          {
            "Authorization": []
          }
        ],
        "summary": "详情",
        "tags": [
          "账本-账单"
        ]
      },
      "delete": {
        "operationId": "BillController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "数据id",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer JWT Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "path": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "data": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          },
          {
            "Authorization": []
          }
        ],
        "summary": "删除",
        "tags": [
          "账本-账单"
        ]
      }
    },
...
  },
  "info": {
    "title": "ledger Example",
    "description": "ledger 接口文档",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "Authorization": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "请输入 JWT Token（示例：eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.xxx）"
      }
    },
    "schemas": {
...
      "CreateBillRes": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "id",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "账单名称",
            "example": "账单名称"
          },
          "amount": {
            "type": "number",
            "description": "账单金额",
            "example": 100
          },
          "classifyId": {
            "type": "number",
            "description": "账单类型id",
            "example": 1
          },
          "userId": {
            "type": "number",
            "description": "创建人id",
            "example": 1
          },
          "billDate": {
            "type": "string",
            "description": "记账日期",
            "example": "2025/12/27"
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间",
            "example": "2025/12/27 20:59:01"
          },
          "updatedAt": {
            "type": "string",
            "description": "更新时间",
            "example": "2025/12/27 20:59:01"
          }
        },
        "required": [
          "id",
          "name",
          "amount",
          "classifyId",
          "userId",
          "billDate",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateBillDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "账单名称",
            "example": "账单名称"
          },
          "amount": {
            "type": "number",
            "description": "账单金额",
            "example": 100
          },
          "classifyId": {
            "type": "number",
            "description": "账单类型id",
            "example": 100
          },
          "ledgerId": {
            "type": "number",
            "description": "账本id",
            "example": 100
          },
          "consumerId": {
            "type": "number",
            "description": "消费人id",
            "example": 1
          },
          "billDate": {
            "type": "string",
            "example": "2025-01-01",
            "description": "记账日期"
          }
        },
        "required": [
          "amount",
          "classifyId",
          "ledgerId",
          "billDate"
        ]
      },
      "UpdateBillDto": {
        "type": "object",
        "properties": {}
      },
      "ClassifyRes": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "应用id",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "应用名称",
            "example": "应用名称"
          },
          "userId": {
            "type": "number",
            "description": "创建人id",
            "example": 1
          },
          "ledgerId": {
            "type": "number",
            "description": "账本id",
            "example": 1
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间",
            "example": "2025/12/27 20:59:01"
          },
          "updatedAt": {
            "type": "string",
            "description": "更新时间",
            "example": "2025/12/27 20:59:01"
          }
        },
        "required": [
          "id",
          "name",
          "userId",
          "ledgerId",
          "createdAt",
          "updatedAt"
        ]
      },
      "LedgerRes": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "id",
            "example": 1
          },
          "userId": {
            "type": "number",
            "description": "创建人id",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "账单名称",
            "example": "账单名称"
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间",
            "example": "2025/12/27 20:59:01"
          },
          "updatedAt": {
            "type": "string",
            "description": "更新时间",
            "example": "2025/12/27 20:59:01"
          }
        },
        "required": [
          "id",
          "userId",
          "name",
          "createdAt",
          "updatedAt"
        ]
      },
      "BillJoinRes": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "id",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "账单名称",
            "example": "账单名称"
          },
          "amount": {
            "type": "number",
            "description": "账单金额",
            "example": 100
          },
          "classifyId": {
            "type": "number",
            "description": "账单类型id",
            "example": 1
          },
          "userId": {
            "type": "number",
            "description": "创建人id",
            "example": 1
          },
          "billDate": {
            "type": "string",
            "description": "记账日期",
            "example": "2025/12/27"
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间",
            "example": "2025/12/27 20:59:01"
          },
          "updatedAt": {
            "type": "string",
            "description": "更新时间",
            "example": "2025/12/27 20:59:01"
          },
          "classify": {
            "description": "分类信息",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClassifyRes"
            }
          },
          "ledger": {
            "description": "账本信息",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LedgerRes"
            }
          }
        },
        "required": [
          "id",
          "name",
          "amount",
          "classifyId",
          "userId",
          "billDate",
          "createdAt",
          "updatedAt",
          "classify",
          "ledger"
        ]
      },
      "ListBillRes": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "description": "页码"
          },
          "size": {
            "type": "number",
            "description": "每页数量"
          },
          "total": {
            "type": "number",
            "description": "数据总数"
          },
          "list": {
            "description": "bill信息",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BillJoinRes"
            }
          }
        },
        "required": [
          "page",
          "size",
          "total",
          "list"
        ]
      },
      "DetailBillRes": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "id",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "账单名称",
            "example": "账单名称"
          },
          "amount": {
            "type": "number",
            "description": "账单金额",
            "example": 100
          },
          "classifyId": {
            "type": "number",
            "description": "账单类型id",
            "example": 1
          },
          "userId": {
            "type": "number",
            "description": "创建人id",
            "example": 1
          },
          "billDate": {
            "type": "string",
            "description": "记账日期",
            "example": "2025/12/27"
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间",
            "example": "2025/12/27 20:59:01"
          },
          "updatedAt": {
            "type": "string",
            "description": "更新时间",
            "example": "2025/12/27 20:59:01"
          }
        },
        "required": [
          "id",
          "name",
          "amount",
          "classifyId",
          "userId",
          "billDate",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}